refactor(resell):独立cv和ocr相关函数
